+++ AppVoor/jsonInfo/welcome.py
@@ -1,31 +1,15 @@
-import json
 import pathlib
-from abc import ABC, abstractmethod
+
+from jsonInfo.json_to_data import ABCJson
+from jsonInfo.random_generator import get_random_number_range_int
 
 
-class ABCJson(ABC):
+class WelcomeMessage(ABCJson):
 
-    def __init__(self, file_name: str):
-        self._file_name = file_name
-        self._data = None
-
-    def _load_file(self, file_name: str):
-        file = pathlib.Path(file_name)
-        if file.exists():
-            with open(file_name, 'r', encoding="utf-8") as f:
-                self._data = json.load(f)
-        raise FileNotFoundError
-
-    @abstractmethod
-    def get_data(self, index: int) -> tuple:
-        pass
-
-
-class WelcomeMessenger(ABCJson):
-
-    def __init__(self, file_name="welcomeMessage.json"):
-        super().__init__(file_name)
-        self._load_file(self._file_name)
+    def __init__(self):
+        file_name = "welcomeMessage.json"
+        super().__init__()
+        self.load_file(file_name)
 
     def get_data(self, index: int) -> tuple:
         if (index < len(self._data)) and (index >= 0):
@@ -33,3 +17,17 @@ class WelcomeMessenger(ABCJson):
             quote = self._data[index]["Quote"]
             return author, quote
         raise IndexError
+
+
+class WelcomeMessenger:
+
+    def __init__(self):
+        self._json_message = WelcomeMessage()
+        random_start = 0
+        random_end = len(self._json_message.get_file_data())
+        self._random_index = get_random_number_range_int(random_start, random_end, 1)
+
+    def get_welcome_message(self) -> str:
+        author, quote = self._json_message.get_data(self._random_index)
+        message = quote + "." + " " + author
+        return message
